# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Digital Credential Common Service API
  version: 1.0.0
  description: |
    Multi-tenant API providing a format-agnostic abstraction layer for Digital Credential operations.
    Normalizes issue/verify/hold patterns across credential formats (AnonCreds, SD-JWT, mDL, W3C VC) and
    back-end agents (Traction/ACA-Py, Credo Agent Service), exposing a single homogeneous REST API.

    This API is organized into three distinct surfaces, each targeting a different audience:

    | Surface | Audience | Auth |
    |---|---|---|
    | **Platform Administration** | Service operators | `platform-admin` role |
    | **Tenant Management** | Tenant administrators | `tenants:admin` scope |
    | **Digital Credential Operations** | Application developers / API consumers | Domain scopes (`credentials:offer`, `credentials:verify`, etc.) |

    ## Authentication
    All endpoints (except health checks and public discovery) require a Bearer JWT issued by the app's
    OIDC provider at `/oidc/token`. Tokens are obtained via:
    - **User login**: Authorization Code + PKCE flow (federated to Keycloak)
    - **Service-to-service**: `client_credentials` grant

    ## Async Operations
    Credential operations that require external party action return `202 Accepted` with an Operation
    resource. Poll `GET /operations/:id` or register webhooks for completion notifications.

    ## Pagination
...

    ## Rate Limiting
...

    ## Idempotency
...

    ## Error Responses
...

...

...

...
servers:
...

...
security:
  - BearerAuth: []
paths:
...
  /api/v1/tenants/{tenantId}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      operationId: getTenant
      summary: Get tenant details
      tags: [Tenants]
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTenant
      summary: Update tenant
      tags: [Tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantRequest"
      responses:
        "200":
          description: Tenant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTenant
      summary: Soft-delete tenant (platform-admin only)
      tags: [Tenants]
      responses:
        "204":
          description: Tenant deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/tenants/{tenantId}/users:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    post:
      operationId: inviteUser
      summary: Invite a user to the tenant
      tags: [Tenant Users]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteUserRequest"
      responses:
        "201":
          description: User invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      operationId: listTenantUsers
      summary: List tenant members
      tags: [Tenant Users]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/TenantUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/tenants/{tenantId}/users/{userId}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: updateTenantUser
      summary: Update user role
      tags: [Tenant Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                display_name:
                  type: string
                role:
                  $ref: "#/components/schemas/TenantRole"
                status:
                  type: string
                  enum: [active, invited, disabled]
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Cannot change the role of the tenant's last owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: removeTenantUser
      summary: Remove user from tenant
      tags: [Tenant Users]
      responses:
        "204":
          description: User removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Cannot remove last owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Connectors ──────────────────────────────────────────────────────
...
  /api/v1/tenants/{tenantId}/roles:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      operationId: listTenantRoles
      summary: List a tenant's effective role scope mappings
      description: |
        Each role reports whether its scopes come from the platform default or
        from a tenant override, so a client can show what has been customised
        and offer a reset.
      tags: [Tenant Settings]
      responses:
        "200":
          description: Effective roles and their scopes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleMappingList"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/tenants/{tenantId}/roles/{role}/scopes:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - name: role
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/TenantRole"
    patch:
      operationId: updateTenantRoleScopes
      summary: Customize role scope mappings for a tenant
      description: |
        Overrides the default role→scope mapping for one role in one tenant.
        The body carries the complete scope list, not a delta.

        Role hierarchy is enforced across the whole mapping, not just the role
        being edited: no role may hold a scope the role above it lacks. The
        `owner` role cannot be modified, and `tenants:admin` cannot be granted
        to any other role.

        **Removing a scope logs out every active user holding this role in the
        tenant.** The `scope` claim is fixed by the OIDC grant created at login
        and is not refreshed when a refresh token rotates, so revoking the
        sessions is the only way to make the change take effect before the
        existing tokens expire. Adding scopes has no such effect and applies at
        the affected users' next login.

        Requires the `tenants:admin` scope.
      tags: [Tenant Settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scopes]
              properties:
                scopes:
                  type: array
                  items:
                    type: string
                  description: Complete replacement scope list for this role in this tenant
      responses:
        "200":
          description: Role scopes updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleScopesResponse"
        "400":
          description: Unknown scope, hierarchy violation, immutable role, or non-assignable scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: resetTenantRoleScopes
      summary: Reset a role to the platform's default scopes
      description: |
        Removes the tenant override so the role inherits the platform default
        again. Idempotent: resetting a role that has no override succeeds and
        returns the default mapping.

        A reset that narrows the role logs out its active users, for the same
        reason a narrowing `PATCH` does.

        Requires the `tenants:admin` scope.
      tags: [Tenant Settings]
      responses:
        "200":
          description: Override removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleScopesResponse"
        "400":
          description: Immutable role or resulting hierarchy violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: App-issued JWT obtained via /oidc/token (client_credentials or authorization_code grant)
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant UUID
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque pagination cursor from a previous response
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema:
        type: string
        format: uuid
      description: UUIDv4 idempotency key to prevent duplicate creation
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: "Tenant not found"
              request_id: "req-uuid"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INSUFFICIENT_SCOPE
              message: "Token missing required scope: credentials:offer"
              request_id: "req-uuid"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: DUPLICATE_RESOURCE
              message: "Tenant with slug 'acme-corp' already exists"
              request_id: "req-uuid"
...
    Unauthorized:
      description: Authentication required or token invalid/expired
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: 'Bearer error="invalid_token", error_description="Token has expired"'
          description: RFC-compliant challenge header
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: AUTHENTICATION_REQUIRED
              message: "Bearer token is missing, expired, or invalid"
              request_id: "req-uuid"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  expected:
                    type: string
                  actual:
                    type: string
                  message:
                    type: string
            request_id:
              type: string
              format: uuid
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            next_cursor:
              type: string
              nullable: true
            has_more:
              type: boolean
            total_count:
              type: integer
              description: Optional total count (may be omitted for performance)
...
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TenantStatus"
        config:
          $ref: "#/components/schemas/TenantConfig"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deactivated_at:
          type: string
          format: date-time
          nullable: true
          description: |
            Set when the tenant is deactivated; cleared on reactivation. Null otherwise. Used to
            measure the 90-day data retention window.
    TenantStatus:
      type: string
      enum:
        [active, pending_approval, rejected, suspended, deactivated]
    TenantConfig:
      type: object
      properties:
        allowed_formats:
          type: array
          items:
            $ref: "#/components/schemas/CredentialFormat"
        default_connector:
          type: string
          format: uuid
          nullable: true
        features:
          type: object
          description: Feature flags (key-value)
        rate_limits:
          type: object
          readOnly: true
          description: |
            Active per-tenant rate limit overrides. Populated when a platform-admin
            has called `PATCH /usage/limits`; absent when platform defaults apply.
          properties:
            api_calls_per_minute:
              type: integer
              nullable: true
            credentials_per_day:
              type: integer
              nullable: true
            storage_mb:
              type: integer
              nullable: true
            expires_at:
              type: string
              format: date-time
              nullable: true
              description: When this override expires (null = permanent)
        operation_ttl:
          type: object
          properties:
            completed_viewed:
              type: string
              example: "1h"
            completed_unviewed:
              type: string
              example: "72h"
            failed_viewed:
              type: string
              example: "24h"
            failed_unviewed:
              type: string
              example: "7d"
            pending_stale:
              type: string
              example: "24h"
...
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
          maxLength: 500

    # ─── Tenant User ─────────────────────────────────────────────────
    TenantUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/TenantRole"
        status:
          type: string
          enum: [active, invited, disabled]
        created_at:
          type: string
          format: date-time
    TenantRole:
      type: string
      enum: [owner, admin, member, readonly]
    RoleScopeSource:
      type: string
      description: |
        `override` when the tenant has customised this role, `default` when it
        inherits the platform mapping.
      enum: [default, override]
    RoleMapping:
      type: object
      properties:
        name:
          $ref: "#/components/schemas/TenantRole"
        scopes:
          type: array
          items:
            type: string
        source:
          $ref: "#/components/schemas/RoleScopeSource"
    RoleMappingList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RoleMapping"
    RoleScopesResponse:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/TenantRole"
        scopes:
          type: array
          items:
            type: string
        source:
          $ref: "#/components/schemas/RoleScopeSource"
        revoked_record_count:
          type: integer
          description: |
            OIDC records deleted because the change removed scopes. Zero when
            the change only widened the role.
    InviteUserRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/TenantRole"

    # ─── Connector ───────────────────────────────────────────────────
...
    CredentialFormat:
      type: string
      enum: [anoncreds, sd-jwt, mdl, w3c-vc]
...